Use unique ids for concurrent imports. Fixes #1.
It can happen that a client opens multiple concurrent sessions at
exactly the same time (time resoultion in log file is seconds). This
led to same ids for different imports. As ImportTrigger
checks based on an import id if an import already run successfully,
only the first import from those with the same id was really executed.